Rename domain.ts to storeUrl.ts
