+++ src/lib/auth/AuthProviders/OAuthAuthenticator/index.ts
@@ -4,7 +4,7 @@ import jwt from 'jsonwebtoken'
 import opn from 'opn'
 import { join } from 'path'
 import { VTEXID } from '../../../../api/clients/IOClients/external/VTEXID'
-import { storeUrl } from '../../../../api/domain'
+import { storeUrl } from '../../../../api/storeUrl'
 import { randomCryptoString } from '../../../utils/randomCryptoString'
 import { spawnUnblockingChildProcess } from '../../../utils/spawnUnblockingChildProcess'
 import { AuthProviderBase } from '../AuthProviderBase'
